move back Value and InputFields into html
